requirements.txt
# Luminate Cookbook Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# Web Framework - FastAPI
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# =============================================================================
# Image Processing
# =============================================================================
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# =============================================================================
# Browser Automation (for Luminate uploads)
# =============================================================================
playwright>=1.40.0; python_version >= "3.8"

# =============================================================================
# HTTP Client
# =============================================================================
requests>=2.28.0
httpx>=0.26.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0

# =============================================================================
# Optional: Google Cloud Storage (for Cloud Run session persistence)
# =============================================================================
# google-cloud-storage>=2.10.0
